remove lib.rs and move all the code to ping.rs
